refactor portfolio activity routes
